Задание module4-task1
